+++ packages/f-checkout/src/constants.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export const VALID_CHECKOUT_TYPES = ['Collection', 'Delivery'];
+export const VALID_CHECKOUT_METHOD = ['Collection', 'Delivery'];